+++ src/main/java/com/codejam/demo/service/ScheduleServiceImpl.java
@@ -1,5 +1,6 @@
 package com.codejam.demo.service;
 
+import com.codejam.demo.exceptions.ResourceNotFoundException;
 import com.codejam.demo.model.Schedule;
 import com.codejam.demo.repository.ScheduleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,16 +19,22 @@ public class ScheduleServiceImpl implements ScheduleService{
 
     @Override
     public Schedule updateSchedule(Schedule schedule) {
+        Schedule dbSchedule = scheduleRepository.findById(schedule.getId()).orElse(null);
+        if(dbSchedule == null) throw new ResourceNotFoundException("Schedule","id",schedule.getId());
         return scheduleRepository.save(schedule);
     }
 
     @Override
     public void deleteSchedule(Integer id) {
+        Schedule dbSchedule = scheduleRepository.findById(id).orElse(null);
+        if(dbSchedule == null) throw new ResourceNotFoundException("Schedule","id",id);
         scheduleRepository.deleteById(id);
     }
 
     @Override
     public Schedule findById(Integer id) {
-        return scheduleRepository.findById(id).orElse(null);
+        Schedule dbSchedule = scheduleRepository.findById(id).orElse(null);
+        if(dbSchedule == null) throw new ResourceNotFoundException("Schedule","id",id);
+        return dbSchedule;
     }
 }
